update module: dubbo-common-limit-starter -> 区分本地限流与redis限流(本地限流：只针对单机器限流，目前集群不支持；redis支持集群限流)

+++ dubbo-common/dubbo-common-limit-starter/src/main/java/com/liyz/dubbo/common/limit/annotation/Limit.java
@@ -20,7 +20,7 @@ public @interface Limit {
      *
      * @return 每秒访问次数
      */
-    double count() default 0.0;
+    long count() default 0;
 
     /**
      * 限流的key：只有在type=customize的时候，该值才会有效
